+++ src/api/session.ts
@@ -3,16 +3,14 @@ import { Operation } from 'express-openapi'
 const env = process.env
 
 export default function (otomi) {
-  const fallbackCluster = env.NODE_ENV !== 'production' ? 'google/dev' : ''
-
   const GET: Operation = [
-    (req, res, next) => {
+    (req, res) => {
       const teamId = req.header('Auth-Group')
       const email = req.header('Auth-User')
       const isAdmin = teamId === 'admin'
       const role = teamId === 'admin' ? 'admin' : 'team'
       const data = {
-        currentClusterId: env.CLUSTER ? env.CLUSTER : fallbackCluster,
+        currentClusterId: env.CLUSTER_ID,
         clusters: otomi.getClusters(),
         core: otomi.getCore(),
         user: { email, teamId, isAdmin, role },