+++ src/api/teams.ts
@@ -2,14 +2,14 @@ import { Operation } from 'express-openapi'
 
 export default function (otomi) {
   const GET: Operation = [
-    (req, res, next) => {
+    (req, res) => {
       const data = otomi.getTeams()
 
       res.status(200).json(data)
     },
   ]
   const POST: Operation = [
-    (req, res, next) => {
+    (req, res) => {
       const data = otomi.createTeam(null, req.body)
       res.status(200).json(data)
     },