+++ src/server.ts
@@ -58,12 +58,5 @@ export default function initApp(otomiStack) {
 
   app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(spec))
 
-  // Serve the static files from the React app
-  app.use('/static/', express.static(path.join(__dirname, './../client/build/static')))
-  // Handles any requests that don't match the ones above
-  app.get('/*', (req, res) => {
-    res.sendFile(path.join(`${__dirname}./../client/build/index.html`))
-  })
-
   return app
 }